// zsphere/apps/dashboard/src/widgets/type.ts
export enum WidgetsType {
  stateMonitor = "state-monitor",
  trendMonitor = "trend-monitor",
  usageStatistics = "usage-statistics",
  topMonitor = "top-monitor",
  recentVisit = "recent-visit",
  userInfo = "user-info",
  alarmInfo = "alarm-info",
  billingStatistics = "billing-statistics",
  quotaUsage = "quota-usage",
  inspection = "inspection",
}